Move AppTheme to its own file
