+++ packages/ubuntu_desktop_installer/lib/main.dart
@@ -7,33 +7,9 @@ import 'package:subiquity_client/subiquity_client.dart';
 import 'package:subiquity_client/subiquity_server.dart';
 
 import 'app.dart';
+import 'app_theme.dart';
 import 'keyboard_model.dart';
 import 'l10n/app_localizations.dart';
-
-class AppTheme extends ValueNotifier<ThemeMode> {
-  AppTheme(this._settings) : super(ThemeMode.system);
-  final GSettings _settings;
-
-  void apply(Brightness brightness) {
-    switch (brightness) {
-      case Brightness.dark:
-        value = ThemeMode.dark;
-        _settings.setValue('gtk-theme', 'Yaru-dark');
-        break;
-      case Brightness.light:
-        value = ThemeMode.light;
-        _settings.setValue('gtk-theme', 'Yaru');
-        break;
-    }
-    _settings.sync();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _settings.dispose();
-  }
-}
 
 Future<void> main() async {
   final themeSettings = GSettings(schemaId: 'org.gnome.desktop.interface');
